+++ src/config.ts
@@ -343,7 +343,7 @@ async function bundleConfigFile(fileName: string, isESM: boolean): Promise<{ cod
     absWorkingDir: process.cwd(),
     entryPoints: [fileName],
     write: false,
-    target: ['node18'],
+    target: ['node20'],
     platform: 'node',
     bundle: true,
     format: isESM ? 'esm' : 'cjs',
@@ -417,6 +417,7 @@ async function loadConfigFormBundledFile(
     } finally {
       try {
         fs.unlinkSync(fileNameTmp)
+        // eslint-disable-next-line no-empty
       } catch {}
     }
   } else {