+++ src/index.ts
@@ -1,4 +1,4 @@
-export { type LogLevel, createLogger, mergeConfig, splitVendorChunkPlugin, splitVendorChunk } from 'vite'
+export { type LogLevel, createLogger, mergeConfig } from 'vite'
 export * from './config'
 export { createServer } from './server'
 export { build } from './build'